% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute_CGI.R
\name{sub_CGI}
\alias{sub_CGI}
\title{Utilily for substitute_CGI}
\usage{
sub_CGI(
  seqs,
  orientation,
  seq_to_add,
  offset,
  direction = c("upstream", "downstream")
)
}
\arguments{
\item{seqs}{A Biostrings sequence to perform operations on}

\item{orientation}{The orientation of the sequence, "+", or "-".}

\item{seq_to_add}{The string to be added}

\item{offset}{How many bps from the center of the sequence should the sequence be substituted.}

\item{direction}{Which direction the substitutions needs to happen. Valid inputs are "upstream" or "downstream".
This is affected by the orientation}
}
\value{
A DNAString object
}
\description{
Utilily for substitute_CGI
}
