% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute_CGI.R
\name{substitute_CGI}
\alias{substitute_CGI}
\title{Add a user provided CG Island anywhere in the sequence. It operates on a sequence and its corresponding orientation}
\usage{
substitute_CGI(
  seqs,
  orientation,
  seq_to_add,
  offset,
  direction = c("upstream", "downstream")
)
}
\arguments{
\item{seqs}{A list of Biostrings sequences to perform operations on}

\item{orientation}{A vector containing the orientation, "+", or "-" of the sequence.}

\item{seq_to_add}{The string to be added}

\item{offset}{How many bps from the center of the sequence should the sequence be added}

\item{direction}{Which direction the substitutions needs to happen. Valid inputs are "upstream" or "downstream".
This is affected by the orientation}
}
\value{
A DNAStringSet object
}
\description{
Add a user provided CG Island anywhere in the sequence. It operates on a sequence and its corresponding orientation
}
\examples{
random_seq <- Biostrings::DNAStringSet("ATCGTATCGATCGATCGATCATCGATCGATCGATCAG")
OR <- c("+")
substitute_CGI(seqs=random_seq, orientation=OR, seq_to_add="CGCG", offset=10, direction="upstream")
}
